+++ src/kio/schema/controlled_shutdown/request/v2.py
@@ -6,6 +6,7 @@ from dataclasses import field
 from typing import ClassVar
 
 from kio.schema.entity import BrokerId
+from kio.schema.primitive import i64
 
 
 @dataclass(frozen=True, slots=True, kw_only=True)
@@ -13,5 +14,5 @@ class ControlledShutdownRequest:
     __flexible__: ClassVar[bool] = False
     broker_id: BrokerId = field(metadata={"kafka_type": "int32"})
     """The id of the broker for which controlled shutdown has been requested."""
-    broker_epoch: int = field(metadata={"kafka_type": "int64"}, default=-1)
+    broker_epoch: i64 = field(metadata={"kafka_type": "int64"}, default=i64(-1))
     """The broker epoch."""
