+++ src/kio/schema/create_acls/request/v1.py
@@ -5,23 +5,25 @@ from dataclasses import dataclass
 from dataclasses import field
 from typing import ClassVar
 
+from kio.schema.primitive import i8
+
 
 @dataclass(frozen=True, slots=True, kw_only=True)
 class AclCreation:
     __flexible__: ClassVar[bool] = False
-    resource_type: int = field(metadata={"kafka_type": "int8"})
+    resource_type: i8 = field(metadata={"kafka_type": "int8"})
     """The type of the resource."""
     resource_name: str = field(metadata={"kafka_type": "string"})
     """The resource name for the ACL."""
-    resource_pattern_type: int = field(metadata={"kafka_type": "int8"}, default=3)
+    resource_pattern_type: i8 = field(metadata={"kafka_type": "int8"}, default=i8(3))
     """The pattern type for the ACL."""
     principal: str = field(metadata={"kafka_type": "string"})
     """The principal for the ACL."""
     host: str = field(metadata={"kafka_type": "string"})
     """The host for the ACL."""
-    operation: int = field(metadata={"kafka_type": "int8"})
+    operation: i8 = field(metadata={"kafka_type": "int8"})
     """The operation type for the ACL (read, write, etc.)."""
-    permission_type: int = field(metadata={"kafka_type": "int8"})
+    permission_type: i8 = field(metadata={"kafka_type": "int8"})
     """The permission type for the ACL (allow, deny, etc.)."""
