+++ src/kio/schema/describe_acls/response/v2.py
@@ -5,6 +5,10 @@ from dataclasses import dataclass
 from dataclasses import field
 from typing import ClassVar
 
+from kio.schema.primitive import i8
+from kio.schema.primitive import i16
+from kio.schema.primitive import i32
+
 
 @dataclass(frozen=True, slots=True, kw_only=True)
 class AclDescription:
@@ -13,20 +17,20 @@ class AclDescription:
     """The ACL principal."""
     host: str = field(metadata={"kafka_type": "string"})
     """The ACL host."""
-    operation: int = field(metadata={"kafka_type": "int8"})
+    operation: i8 = field(metadata={"kafka_type": "int8"})
     """The ACL operation."""
-    permission_type: int = field(metadata={"kafka_type": "int8"})
+    permission_type: i8 = field(metadata={"kafka_type": "int8"})
     """The ACL permission type."""
 
 
 @dataclass(frozen=True, slots=True, kw_only=True)
 class DescribeAclsResource:
     __flexible__: ClassVar[bool] = True
-    resource_type: int = field(metadata={"kafka_type": "int8"})
+    resource_type: i8 = field(metadata={"kafka_type": "int8"})
     """The resource type."""
     resource_name: str = field(metadata={"kafka_type": "string"})
     """The resource name."""
-    pattern_type: int = field(metadata={"kafka_type": "int8"}, default=3)
+    pattern_type: i8 = field(metadata={"kafka_type": "int8"}, default=i8(3))
     """The resource pattern type."""
     acls: tuple[AclDescription, ...]
     """The ACLs."""
@@ -35,9 +39,9 @@ class DescribeAclsResource:
 @dataclass(frozen=True, slots=True, kw_only=True)
 class DescribeAclsResponse:
     __flexible__: ClassVar[bool] = True
-    throttle_time_ms: int = field(metadata={"kafka_type": "int32"})
+    throttle_time_ms: i32 = field(metadata={"kafka_type": "int32"})
     """The duration in milliseconds for which the request was throttled due to a quota violation, or zero if the request did not violate any quota."""
-    error_code: int = field(metadata={"kafka_type": "int16"})
+    error_code: i16 = field(metadata={"kafka_type": "int16"})
     """The error code, or 0 if there was no error."""
     error_message: str | None = field(metadata={"kafka_type": "string"})
     """The error message, or null if there was no error."""
