+++ src/kio/schema/describe_configs/request/v0.py
@@ -5,11 +5,13 @@ from dataclasses import dataclass
 from dataclasses import field
 from typing import ClassVar
 
+from kio.schema.primitive import i8
+
 
 @dataclass(frozen=True, slots=True, kw_only=True)
 class DescribeConfigsResource:
     __flexible__: ClassVar[bool] = False
-    resource_type: int = field(metadata={"kafka_type": "int8"})
+    resource_type: i8 = field(metadata={"kafka_type": "int8"})
     """The resource type."""
     resource_name: str = field(metadata={"kafka_type": "string"})
     """The resource name."""
