+++ src/kio/schema/describe_delegation_token/response/v2.py
@@ -5,6 +5,10 @@ from dataclasses import dataclass
 from dataclasses import field
 from typing import ClassVar
 
+from kio.schema.primitive import i16
+from kio.schema.primitive import i32
+from kio.schema.primitive import i64
+
 
 @dataclass(frozen=True, slots=True, kw_only=True)
 class DescribedDelegationTokenRenewer:
@@ -22,11 +26,11 @@ class DescribedDelegationToken:
     """The token principal type."""
     principal_name: str = field(metadata={"kafka_type": "string"})
     """The token principal name."""
-    issue_timestamp: int = field(metadata={"kafka_type": "int64"})
+    issue_timestamp: i64 = field(metadata={"kafka_type": "int64"})
     """The token issue timestamp in milliseconds."""
-    expiry_timestamp: int = field(metadata={"kafka_type": "int64"})
+    expiry_timestamp: i64 = field(metadata={"kafka_type": "int64"})
     """The token expiry timestamp in milliseconds."""
-    max_timestamp: int = field(metadata={"kafka_type": "int64"})
+    max_timestamp: i64 = field(metadata={"kafka_type": "int64"})
     """The token maximum timestamp length in milliseconds."""
     token_id: str = field(metadata={"kafka_type": "string"})
     """The token ID."""
@@ -39,9 +43,9 @@ class DescribedDelegationToken:
 @dataclass(frozen=True, slots=True, kw_only=True)
 class DescribeDelegationTokenResponse:
     __flexible__: ClassVar[bool] = True
-    error_code: int = field(metadata={"kafka_type": "int16"})
+    error_code: i16 = field(metadata={"kafka_type": "int16"})
     """The error code, or 0 if there was no error."""
     tokens: tuple[DescribedDelegationToken, ...]
     """The tokens."""
-    throttle_time_ms: int = field(metadata={"kafka_type": "int32"})
+    throttle_time_ms: i32 = field(metadata={"kafka_type": "int32"})
     """The duration in milliseconds for which the request was throttled due to a quota violation, or zero if the request did not violate any quota."""
